김예영: NumberRender 업데이트

+++ APIApp/GameEngineCore/GameEngineRender.cpp
@@ -119,7 +119,7 @@ void GameEngineRender::TextRender(float _DeltaTime)
 
 	float4 RenderPos = GetActorPlusPos() - CameraPos;
 
-	TextOutA(GameEngineWindow::GetDoubleBufferImage()->GetImageDC(), RenderPos.ix(), RenderPos.iy(), RenderText.c_str(), RenderText.size());
+	TextOutA(GameEngineWindow::GetDoubleBufferImage()->GetImageDC(), RenderPos.ix(), RenderPos.iy(), RenderText.c_str(), static_cast<int>(RenderText.size()));
 
 	return;
 }
